refactor: break monolithic api.ts into 7 domain routers
Replace the 1041-line routes/api.ts with modular domain routers:
- routes/auth.ts (137 lines) — register, signup-status
- routes/runners.ts (441 lines) — spawn, restart, stop, terminal, skills, files, git
- routes/sessions.ts (66 lines) — list, pin/unpin
- routes/attachments.ts (115 lines) — upload, download
- routes/chat.ts (75 lines) — chat completion, model listing
- routes/push.ts (89 lines) — VAPID key, subscribe/unsubscribe, events
- routes/settings.ts (38 lines) — hidden models

Infrastructure:
- routes/types.ts — shared RouteHandler type
- routes/utils.ts — extracted parseJsonArray, pickRunnerIdLeastLoaded,
  pickRunnerIdForCwd, mintEphemeralApiKey
- routes/index.ts — central dispatcher chains routers sequentially
- handler.ts updated to import from routes/index.ts

100% backward compatible — all endpoints, status codes, and behavior
unchanged. All 115 existing tests pass.

Completes tasks 001-005 of epic/api-refactor-routes.
