+++ .gitignore
@@ -9,3 +9,4 @@ docs-homepage.png
 docs-getting-started.png
 *.tsbuildinfo
 auth.db
+.Jules/
